Final Changes for tonight
Patched up the process by which we will check to see if data's inputed.
Right now, if you fill out the form, it sends it to another page to
check for data, then responds success or failure.  Allows us to remark
on whether data was recorded, or whether there was a problem.